+++ src/main/java/proyect/cema/Services/Models/ProductDTO.java
@@ -4,7 +4,7 @@ package proyect.cema.Services.Models;
 public class ProductDTO {
     private Long id;
     private String nombre_producto;
-    private Long precio;
+    private double precio;
     private String talla;
     private String imagenes;
     private String color;
@@ -16,7 +16,7 @@ public class ProductDTO {
     public ProductDTO() { }
 
 
-   public ProductDTO(String nombre_producto, Long precio, String talla, String imagenes, String color, String estado,
+   public ProductDTO(String nombre_producto, double precio, String talla, String imagenes, String color, String estado,
            String marca, String sexo) {
        this.nombre_producto = nombre_producto;
        this.precio = precio;
@@ -49,12 +49,15 @@ public void setNombre_producto(String nombre_producto) {
 }
 
 
-public Long getPrecio() {
+
+
+
+public double getPrecio() {
     return precio;
 }
 
 
-public void setPrecio(Long precio) {
+public void setPrecio(double precio) {
     this.precio = precio;
 }
